# CUDA: faster softmax via shared memory + fp16 math