This PR adds performance optimizations for CUDA softmax calculation. According to profiling data the softmax kernel has become 2.4 times faster. However, because the softmax kernel only takes up a small percentage of the runtime the overall impact on performance is small:

| Model         | GPU                     | test   |   t/s master |   t/s PR |   Speedup |
|---------------|-------------------------|--------|--------------|----------|-----------|
| llama 7B Q4_0 | NVIDIA GeForce RTX 3090 | pp512  |      3462.88 |  3527.40 |      1.02 |
| llama 7B Q4_0 | NVIDIA GeForce RTX 3090 | tg128  |       136.90 |   136.73 |      1.00 |
| llama 7B Q4_0 | Tesla P40               | pp512  |       894.99 |   908.75 |      1.02 |
| llama 7B Q4_0 | Tesla P40               | tg128  |        56.67 |    56.84 |      1.00 |
| llama 7B Q4_0 | AMD Radeon RX 6800      | pp512  |      1241.66 |  1248.01 |      1.01 |
| llama 7B Q4_0 | AMD Radeon RX 6800      | tg128  |        74.41 |    74.31 |      1.00 |

The performance optimizations include:

* Using shared memory to avoid loading the data multiple times from global memory.
* Template parameter to optionally inform the compiler of the number of columns and the block size. If both are specified the compiler can unroll loops, and avoid checks for correctness. Aside from a generic kernel that can handle any number of columns fast variants for powers of 2 are compiled. These are the sizes used for prompt processing where softmax (on master) is taking up a relatively high percentage of the runtime compared to token generation.
* On NVIDIA cards that are Volta or newer FP16 arithmetic instead of FP32 arithmetic is used. This is both faster and reduces the amount of shared memory required. Perplexity for LLaMA 2 q4_0 changes from $5.9634 \pm 0.03349$ to $5.9629 \pm 0.03349$. For AMD FP16 arithmetic is disabled because the HIP documentation does not seem to contain an equivalent for `__hmax2`.