@slaren 2024-01-02
`test-backend-ops` fails with this change:
```
$ tests/test-backend-ops -b CUDA0 -o SOFT_MAX                                                                                                            Testing 2 backends

Backend 1/2 (CPU)
  Skipping
Backend 2/2 (CUDA0)
ggml_init_cublas: GGML_CUDA_FORCE_MMQ:   no
ggml_init_cublas: CUDA_USE_TENSOR_CORES: yes
ggml_init_cublas: found 1 CUDA devices:
  Device 0: NVIDIA GeForce RTX 3090 Ti, compute capability 8.6, VMM: yes
  Backend name: CUDA
  SOFT_MAX(type=f32,ne=[10,10,10,10]): [SOFT_MAX] NMSE = 4.772921 sentinel mismatch: sent_2 FAIL
  876/877 tests passed
  Backend CUDA: FAIL

1/2 backends passed
**FAIL**
```

The sentinel mismatch indicates a buffer overflow.

We need to be careful about downcasting to F16 without the user explicitly requesting this, as some models are susceptible to precision issues. We have had some issues already with F16 matrix multiplication, with Phi2 it causes issue in the KV cache, and in whisper.cpp it causes degraded quality. I think that stable diffusion is also susceptible to precision issues. Currently we have a function to set the precision of matrix multiplications, but I think that we need a more general solution.