@JohannesGaessler 2024-01-02
>The sentinel mismatch indicates a buffer overflow.

Thank you. I didn't notice with `compute-sanitizer` because this only seems to happen for batch sizes with % 32 != 1.

>We need to be careful about downcasting to F16 without the user explicitly requesting this, as some models are susceptible to precision issues.

Good point. I changed the check to choose simply based on `GGML_CUDA_F16`.

More generally, what is the long-term plan for CUDA FP16 support in ggml? I think to remember that one of your unmerged PRs had more general FP16 support so I was under the impression that it is of at least moderate priority.