@slaren 2024-01-03
And at the moment my priority is to move the CUDA backend implementation fully to the ggml-backend interface, and then do some refactoring of the CUDA backend. Essentially what this means is that partial offloading will be handled entirely by ggml-backend with `ggml_backend_sched`, and most of the code that deals with copying data between the CPU and the GPU in the CUDA backend will be removed. Then we can finally remove most of the globals in ggml-cuda, make it thread safe, add better support for multiple devices and so on.

So adding support for FP16 is not a priority for me at the moment. I had a partial implementation in the [first iteration](https://github.com/ggerganov/llama.cpp/tree/ggml-backends) of the ggml-backend interface, and some of that may still be useful today. The way it worked was that `ggml_context` had a parameter that allowed setting the type of the operations, in that way operations could return FP16 tensors instead of always returning FP32 as it does now. In the CUDA backend side, all kernels were made templates that could use any floating point type. I think there were some mistakes there, for instance there is no reason to use FP16 arithmetic unless using SIMD with `half2`, but overall I think the design could still work.