@JohannesGaessler 2024-01-03
>for instance there is no reason to use FP16 arithmetic unless using SIMD with half2

I would say this depends on the computing task. Currently the focus is on inference where the static weights take up the most memory but e.g. for training neural networks you need a lot of memory to store gradients. In such scenarios FP16 would be useful simply to save space compared to FP32. But for the current state of the software I definitely agree.