@kalomaze 2024-01-03
Are there any plans to move sampling logic to the GPU as well? As it currently stands, multiple sampler passes (for Top K, Temperature, Min P, repetition penalty, etc) are running on CPU without any threading. The softmax is just one operation that is currently happening on CPU. Certain more complex samplers are also not trivially inexpensive on CPU due to large vocabulary sizes, especially the custom Grammar sampler.