@JohannesGaessler 2024-01-03
I'm not overly familiar with the sampling code but if I remember correctly it has a lot of branching. GPUs are only faster than CPUs for code where you can apply the same instructions to multiple data streams at the same time. So I'm not sure that sampling on the GPU would even be any faster.