@slaren 2024-01-03
Btw, `test-backend-ops` also has a mode for testing performance, just add `perf` to the command line. I get these results:

New test cases:
```C++
test_cases.emplace_back(new test_soft_max(GGML_TYPE_F32, {4096, 1, 1, 1}));
test_cases.emplace_back(new test_soft_max(GGML_TYPE_F32, {1024, 512, 1, 1}));
test_cases.emplace_back(new test_soft_max(GGML_TYPE_F32, {4096, 512, 1, 1}));
```

master:
```
SOFT_MAX(type=f32,ne=[10,10,10,10]):                           8192 runs -     5.69 us/run -       78 kB/run -   13.09 GB/s
SOFT_MAX(type=f32,ne=[4096,1,1,1]):                            8192 runs -     7.15 us/run -       32 kB/run -    4.27 GB/s
SOFT_MAX(type=f32,ne=[1024,512,1,1]):                          8192 runs -    16.97 us/run -     4096 kB/run -  230.13 GB/s
SOFT_MAX(type=f32,ne=[4096,512,1,1]):                          2049 runs -    24.97 us/run -    16384 kB/run -  625.66 GB/s
```

PR:
```
SOFT_MAX(type=f32,ne=[10,10,10,10]):                           8192 runs -     6.04 us/run -       78 kB/run -   12.34 GB/s
SOFT_MAX(type=f32,ne=[4096,1,1,1]):                            8192 runs -     6.37 us/run -       32 kB/run -    4.79 GB/s
SOFT_MAX(type=f32,ne=[1024,512,1,1]):                          8192 runs -    12.82 us/run -     4096 kB/run -  304.65 GB/s
SOFT_MAX(type=f32,ne=[4096,512,1,1]):                          2049 runs -    23.56 us/run -    16384 kB/run -  663.12 GB/s
```

PR + LLAMA_CUDA_F16:
```
SOFT_MAX(type=f32,ne=[10,10,10,10]):                           8192 runs -     5.51 us/run -       78 kB/run -   13.52 GB/s
SOFT_MAX(type=f32,ne=[4096,1,1,1]):                            8192 runs -     6.18 us/run -       32 kB/run -    4.94 GB/s
SOFT_MAX(type=f32,ne=[1024,512,1,1]):                          8192 runs -     8.99 us/run -     4096 kB/run -  434.42 GB/s
SOFT_MAX(type=f32,ne=[4096,512,1,1]):                          2049 runs -    22.80 us/run -    16384 kB/run -  685.35 GB/s
```


Building with `LLAMA_CUDA_F16` also makes the tests fail, which I guess it is expected, however the error in the first case seems to high:

```
  SOFT_MAX(type=f32,ne=[10,10,10,10]): [SOFT_MAX] NMSE = 4.806375 FAIL
  SOFT_MAX(type=f32,ne=[4096,1,1,1]): [SOFT_MAX] NMSE = 0.000000 FAIL
  SOFT_MAX(type=f32,ne=[1024,512,1,1]): [SOFT_MAX] NMSE = 0.000000 FAIL
  SOFT_MAX(type=f32,ne=[4096,512,1,1]): [SOFT_MAX] NMSE = 0.000000 FAIL
```