@JohannesGaessler 2024-01-03
>however the error in the first case seems too high

Yes, it seems the checks for arbitrary tensor sizes were still not quite correct.

I've also committed the changes I made to `test-backend-ops.cpp` for debugging. The output now looks like this:

```
> ./tests/test-backend-ops -o SOFT_MAX                                
Testing 2 backends

Backend 1/2 (CPU)
  Backend name: CPU
ggml_init_cublas: GGML_CUDA_FORCE_MMQ:   no
ggml_init_cublas: CUDA_USE_TENSOR_CORES: yes
ggml_init_cublas: found 1 CUDA devices:
  Device 0: NVIDIA GeForce RTX 3090, compute capability 8.6, VMM: yes
  SOFT_MAX(type=f32,ne=[10,10,10,10]): OK
  877/877 tests passed
  Backend CPU: OK

Backend 2/2 (CUDA0)
  Backend name: CUDA
  SOFT_MAX(type=f32,ne=[10,10,10,10]): [SOFT_MAX] NMSE = 2.362E-07 > 1.000E-07 FAIL
  876/877 tests passed
  Backend CUDA: FAIL

1/2 backends passed
FAIL
```

I think this is better because it tells you how large some metric is compared to some threshold. "NMSE" to me doesn't mean anything without knowing what the code does so I wouldn't know how good/or bad a value of 4.8 is. But with a comparison to a threshold of 1e-7 it is clear that there is still a bug somewhere.