@slaren 2024-01-04
Recap of the performance that I see with `test-backend-ops`, 3090 Ti, WSL:

| test                                              | master     | PR           | PR F16      | speedup | speedup F16 |
| ------------------------------------------------- | --------:| ---------:   | -----:      | ------: | ----------: |
|  SOFT_MAX(type=f32,ne=[10,10,10,10])              | 20.25 GB/s | 20.28 GB/s   | 22.07 GB/s  | 1.001 | 1.089 |
|  SOFT_MAX(type=f32,ne=[4096,1,1,1])               |  5.37 GB/s |  7.34 GB/s   |  7.51 GB/s  | 1.366 | 1.398 |
|  SOFT_MAX(type=f32,ne=[1024,512,1,1])             | 221.22 GB/s | 313.56 GB/s | 544.22 GB/s | 1.417 | 2.460 |
|  SOFT_MAX(type=f32,ne=[4096,512,1,1])             | 637.94 GB/s | 683.83 GB/s | 685.38 GB/s | 1.071 | 1.074 |
|  SOFT_MAX(type=f32,ne=[4096,16384,1,1])           | 791.93 GB/s | 837.56 GB/s | 837.13 GB/s | 1.057 | 1.057 |

There may still be some issues with F16, I tried some random dimensions and some tests have a very large error:
```
  SOFT_MAX(type=f32,ne=[27068,3,1,1]): [SOFT_MAX] NMSE = 1.000E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[60364,480,1,1]): [SOFT_MAX] NMSE = 2.123E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[40397,67,1,1]): [SOFT_MAX] NMSE = 1.753E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[60604,151,1,1]): [SOFT_MAX] NMSE = 2.240E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[95781,111,1,1]): [SOFT_MAX] NMSE = 6.218E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[91341,24,1,1]): [SOFT_MAX] NMSE = 8.201E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[92999,316,1,1]): [SOFT_MAX] NMSE = 5.111E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[59815,366,1,1]): [SOFT_MAX] NMSE = 2.363E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[85203,225,1,1]): [SOFT_MAX] NMSE = 4.698E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[34963,461,1,1]): [SOFT_MAX] NMSE = 1.000E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[90831,115,1,1]): [SOFT_MAX] NMSE = 5.474E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[69272,34,1,1]): [SOFT_MAX] NMSE = 5.267E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[83132,229,1,1]): [SOFT_MAX] NMSE = 4.389E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[59246,17,1,1]): [SOFT_MAX] NMSE = 3.770E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[81631,48,1,1]): [SOFT_MAX] NMSE = 4.178E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[59536,390,1,1]): [SOFT_MAX] NMSE = 1.860E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[86513,329,1,1]): [SOFT_MAX] NMSE = 4.791E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[64085,127,1,1]): [SOFT_MAX] NMSE = 2.969E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[99567,221,1,1]): [SOFT_MAX] NMSE = 5.428E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[68392,54,1,1]): [SOFT_MAX] NMSE = 2.938E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[73394,387,1,1]): [SOFT_MAX] NMSE = 3.807E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[45188,221,1,1]): [SOFT_MAX] NMSE = 1.728E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[44265,162,1,1]): [SOFT_MAX] NMSE = 1.675E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[86377,410,1,1]): [SOFT_MAX] NMSE = 4.713E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[92523,330,1,1]): [SOFT_MAX] NMSE = 4.958E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[67839,104,1,1]): [SOFT_MAX] NMSE = 3.744E+00 > 1.000E-07 FAIL
  SOFT_MAX(type=f32,ne=[64483,77,1,1]): [SOFT_MAX] NMSE = 2.855E+00 > 1.000E-07 FAIL
```

To generate the tests I used this code:
```C++
    std::random_device rd;
    std::default_random_engine rng(rd());
    std::uniform_int_distribution<> dist1(1, 100000);
    std::uniform_int_distribution<> dist2(1, 512);
    for (int n = 0; n < 100; n++) {
        int64_t ne0 = dist1(rng);
        int64_t ne1 = dist2(rng);
        test_cases.emplace_back(new test_soft_max(GGML_TYPE_F32, {ne0, ne1, 1, 1}));
    }
```

I am not convinced about the scientific notation in the NMSE error, it makes harder to see at a glance what errors are huge. Maybe we could increase the number of decimals printed?